backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==4.4.2
filelock==3.17.0
flask==2.0.1
flask-cors==3.0.10
frozenlist==1.5.0
fsspec==2024.12.0
gradio_client==1.5.4
grpcio==1.69.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
logging==0.4.9.6
moviepy==1.0.3
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.3
PyAudio==0.2.14
pydub==0.25.1
pygame==2.6.1
pyht==0.1.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==13.1
Werkzeug==2.0.1
yarl==1.18.3
